+++ tests/indices/vector_store/test_deeplake.py
@@ -9,7 +9,6 @@ from llama_index.indices.vector_store.base import VectorStoreIndex
 from llama_index.schema import Document, TextNode
 from llama_index.storage.storage_context import StorageContext
 from llama_index.vector_stores import DeepLakeVectorStore
-from llama_index.vector_stores.types import NodeWithEmbedding
 
 try:
     import deeplake
@@ -65,12 +64,10 @@ def test_build_deeplake(
 
     node = nodes[0].node
 
-    result = NodeWithEmbedding(
-        node=node,
-        embedding=[1.0 for i in range(EMBEDDING_DIM)],
-    )
-    results = [result for i in range(NUMBER_OF_DATA)]
-    vector_store.add(results)
+    node_with_embedding = node.copy()
+    node_with_embedding.embedding = [1.0 for i in range(EMBEDDING_DIM)]
+    new_nodes = [node_with_embedding for i in range(NUMBER_OF_DATA)]
+    vector_store.add(new_nodes)
     assert len(vector_store.vectorstore) == 14
 
     ref_doc_id = str(node.ref_doc_id)
@@ -119,14 +116,13 @@ def test_backwards_compatibility() -> None:
     )
     metadatas = [metadata.update({"doc_id": "2"}) for metadata in metadatas]
     node = TextNode(
-        text="test node text", metadata={"key": "value", "doc_id": "1"}, id_="1"
-    )
-    result = NodeWithEmbedding(
-        node=node,
+        text="test node text",
+        metadata={"key": "value", "doc_id": "1"},
+        id_="1",
         embedding=[1.0 for i in range(EMBEDDING_DIM)],
     )
 
-    results = [result for i in range(10)]
+    nodes = [node for i in range(10)]
 
     dataset_path = "local_ds1"
     ds = deeplake.empty(dataset_path)
@@ -150,6 +146,6 @@ def test_backwards_compatibility() -> None:
         verbose=False,
     )
 
-    vectorstore.add(results)
+    vectorstore.add(nodes)
     assert len(vectorstore.vectorstore) == 20
     deeplake.delete(dataset_path)