+++ tests/vector_stores/test_qdrant.py
@@ -10,7 +10,6 @@ except ImportError:
 from llama_index.schema import NodeRelationship, RelatedNodeInfo, TextNode
 from llama_index.vector_stores import QdrantVectorStore
 from llama_index.vector_stores.types import (
-    NodeWithEmbedding,
     VectorStoreQuery,
     MetadataFilters,
     ExactMatchFilter,
@@ -18,41 +17,33 @@ from llama_index.vector_stores.types import (
 
 
 @pytest.fixture
-def node_embeddings() -> List[NodeWithEmbedding]:
+def node_embeddings() -> List[TextNode]:
     return [
-        NodeWithEmbedding(
+        TextNode(
+            text="lorem ipsum",
+            id_="c330d77f-90bd-4c51-9ed2-57d8d693b3b0",
+            relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="test-0")},
+            metadata={
+                "author": "Stephen King",
+                "theme": "Friendship",
+            },
             embedding=[1.0, 0.0],
-            node=TextNode(
-                text="lorem ipsum",
-                id_="c330d77f-90bd-4c51-9ed2-57d8d693b3b0",
-                relationships={
-                    NodeRelationship.SOURCE: RelatedNodeInfo(node_id="test-0")
-                },
-                metadata={
-                    "author": "Stephen King",
-                    "theme": "Friendship",
-                },
-            ),
         ),
-        NodeWithEmbedding(
+        TextNode(
+            text="lorem ipsum",
+            id_="c3d1e1dd-8fb4-4b8f-b7ea-7fa96038d39d",
+            relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="test-1")},
+            metadata={
+                "director": "Francis Ford Coppola",
+                "theme": "Mafia",
+            },
             embedding=[0.0, 1.0],
-            node=TextNode(
-                text="lorem ipsum",
-                id_="c3d1e1dd-8fb4-4b8f-b7ea-7fa96038d39d",
-                relationships={
-                    NodeRelationship.SOURCE: RelatedNodeInfo(node_id="test-1")
-                },
-                metadata={
-                    "director": "Francis Ford Coppola",
-                    "theme": "Mafia",
-                },
-            ),
         ),
     ]
 
 
 @pytest.mark.skipif(qdrant_client is None, reason="qdrant-client not installed")
-def test_add_stores_data(node_embeddings: List[NodeWithEmbedding]) -> None:
+def test_add_stores_data(node_embeddings: List[TextNode]) -> None:
     client = qdrant_client.QdrantClient(":memory:")
     qdrant_vector_store = QdrantVectorStore(collection_name="test", client=client)